bake-decouple: move bake source providers out of bun_sourcemap
